+++ pkg/channel/store/memory.go
@@ -796,6 +796,7 @@ func messageFromRecord(id ch.ChannelID, record ch.Record) ch.Message {
 		ServerTimestampMS: record.ServerTimestampMS,
 		SyncOnce:          record.SyncOnce,
 		RedDot:            record.RedDot,
+		Expire:            record.Expire,
 	}
 }